add original codes(test)

+++ train.py
@@ -168,7 +168,7 @@ def setup_training_loop_kwargs(
         spec.ref_gpus = gpus
         res = args.training_set_kwargs.resolution
         # spec.mb = max(min(gpus * min(4096 // res, 32), 64), gpus) # keep gpu memory consumption at bay
-        spec.mb = max(min(gpus * min(4096 // res, 32), 64), gpus) / 2 # keep gpu memory consumption at bay
+        spec.mb = int(max(min(gpus * min(4096 // res, 32), 64), gpus) / 2) # keep gpu memory consumption at bay
         spec.mbstd = min(spec.mb // gpus, 4) # other hyperparams behave more predictably if mbstd group size remains fixed
         spec.fmaps = 1 if res >= 512 else 0.5
         spec.lrate = 0.002 if res >= 1024 else 0.0025
